# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental API",
        "description": "This is the API documentation for the car rental service.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/rentals": {
            "get": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Get all rentals",
                "description": "Display a listing of the rentals.",
                "operationId": "a19748eac97c416a4a02a73fbcdb4700",
                "responses": {
                    "200": {
                        "description": "A list of rentals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Rental"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Create a new rental",
                "description": "Store a newly created rental.",
                "operationId": "7e700bb6638cc9e6a42aa112bf018ba6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "car_id",
                                    "start_date",
                                    "end_date",
                                    "total_price",
                                    "status"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "car_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-03-10"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-03-20"
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "active",
                                            "completed",
                                            "canceled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rental created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rental"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/rentals/{id}": {
            "get": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Get a rental by ID",
                "description": "Display the specified rental.",
                "operationId": "7248fde8ecc49afa506485b780b4cd27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rental ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rental details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rental"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rental not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rental not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Update an existing rental",
                "description": "Update the specified rental.",
                "operationId": "45134c777261098392b52d9f89a0311a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rental ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "car_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-03-10"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-03-20"
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "active",
                                            "completed",
                                            "canceled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rental updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rental"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rental not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rental not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Delete a rental",
                "description": "Remove the specified rental.",
                "operationId": "f130db03982a2f6048b833c6cc977fcf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rental ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rental deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rental deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rental not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rental not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{userId}/rentals/": {
            "get": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Get rentals for a specific user",
                "description": "Get rentals by a specific user.",
                "operationId": "3c46789194c8347cc15fcf4657cdad6a",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of rentals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Rental"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cars/{carId}/rentals": {
            "get": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Get rentals for a specific car",
                "description": "Get rentals for a specific car.",
                "operationId": "c321c6ae981acbc9dce5e8e7cf8eb939",
                "parameters": [
                    {
                        "name": "carId",
                        "in": "path",
                        "description": "Car ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of rentals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Rental"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Car not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Rental": {
                "required": [
                    "user_id",
                    "car_id",
                    "start_date",
                    "end_date",
                    "total_price",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the rental",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID of the user renting the car",
                        "type": "integer",
                        "example": 1
                    },
                    "car_id": {
                        "description": "ID of the rented car",
                        "type": "integer",
                        "example": 1
                    },
                    "start_date": {
                        "description": "Rental start date",
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-10"
                    },
                    "end_date": {
                        "description": "Rental end date",
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-20"
                    },
                    "total_price": {
                        "description": "Total rental price",
                        "type": "number",
                        "format": "float",
                        "example": 250
                    },
                    "status": {
                        "description": "Status of the rental",
                        "type": "string",
                        "enum": [
                            "pending",
                            "active",
                            "completed",
                            "canceled"
                        ],
                        "example": "active"
                    },
                    "created_at": {
                        "description": "Timestamp when the rental was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-10T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the rental was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-10T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}